+++ src/promptflow-core/promptflow/core/_prompty_utils.py
@@ -3,9 +3,7 @@ import json
 import os
 import re
 from dataclasses import asdict
-from os import PathLike
-from pathlib import Path
-from typing import List, Mapping, Union
+from typing import List, Mapping
 
 from promptflow.core._connection import AzureOpenAIConnection, OpenAIConnection, _Connection
 from promptflow.core._errors import (
@@ -13,7 +11,6 @@ from promptflow.core._errors import (
     ChatAPIInvalidRoleError,
     CoreError,
     InvalidOutputKeyError,
-    InvalidSampleError,
     UnknownConnectionType,
 )
 from promptflow.core._model_configuration import ModelConfiguration
@@ -46,27 +43,13 @@ def parse_environment_variable(value):
         return value
 
 
-def load_inputs_from_sample(sample: Union[dict, str, PathLike]):
-    if not sample:
-        return {}
-    elif isinstance(sample, dict):
-        return sample
-    elif isinstance(sample, (str, Path)) and Path(sample).suffix.lower() == ".json":
-        if not Path(sample).exists():
-            raise InvalidSampleError(f"Cannot find sample file {sample}.")
-        with open(sample, "r") as f:
-            return json.load(f)
-    else:
-        raise InvalidSampleError("Only dict and json file are supported as sample in prompty.")
-
-
 def get_connection_by_name(connection_name):
     try:
         from promptflow._sdk._pf_client import PFClient
     except ImportError as ex:
         raise CoreError(f"Please try 'pip install promptflow-devkit' to install dependency, {ex.msg}")
     client = PFClient()
-    connection_obj = client.connections.get(connection_name, with_secrets=True)
+    connection_obj = client.connections._get(connection_name, with_secrets=True)
     connection = connection_obj._to_execution_connection_dict()["value"]
     connection_type = connection_obj.TYPE
     return connection, connection_type