+++ src/promptflow-devkit/promptflow/_sdk/entities/_experiment.py
@@ -101,6 +101,7 @@ class FlowNode(YAMLTranslatableMixin):
         environment_variables: Optional[Dict[str, str]] = None,
         connections: Optional[Dict[str, Dict]] = None,
         properties: Optional[Dict[str, Any]] = None,
+        init: Optional[dict] = None,
         **kwargs,
     ):
         self.type = ExperimentNodeType.FLOW
@@ -118,6 +119,7 @@ class FlowNode(YAMLTranslatableMixin):
         self.path = path
         # default run name: flow directory name + timestamp
         self.name = name
+        self.init = init or {}
         self._runtime = kwargs.get("runtime", None)
         self._resources = kwargs.get("resources", None)
 
@@ -158,6 +160,8 @@ class CommandNode(YAMLTranslatableMixin):
         environment_variables=None,
         code=None,
         display_name=None,
+        resources=None,
+        identity=None,
         **kwargs,
     ):
         self.type = ExperimentNodeType.COMMAND
@@ -168,6 +172,8 @@ class CommandNode(YAMLTranslatableMixin):
         self.inputs = inputs or {}
         self.outputs = outputs or {}
         self.runtime = runtime
+        self.resources = resources
+        self.identity = identity
         self.environment_variables = environment_variables or {}
 
     @classmethod