+++ src/promptflow/tests/executor/unittests/_utils/test_tool_utils.py
@@ -5,7 +5,7 @@ import pytest
 
 from promptflow._core._errors import DuplicateToolMappingError
 from promptflow._utils.tool_utils import (
-    DynamicListError,
+    RetrieveToolFuncResultError,
     ListFunctionResponseError,
     RetrieveToolFuncResultValidationError,
     _find_deprecated_tools,
@@ -323,7 +323,7 @@ class TestToolUtils:
     )
     def test_validate_dynamic_list_func_response_type_with_error(self, res, err_msg):
         error_message = (
-            f"Unable to display list of items due to '{err_msg}'. \nPlease contact the tool "
+            f"Unable to retrieve result due to '{err_msg}'. \nPlease contact the tool "
             f"author/support team for troubleshooting assistance."
         )
         with pytest.raises(ListFunctionResponseError, match=error_message):
@@ -342,8 +342,8 @@ class TestToolUtils:
     def test_load_function_from_function_path_with_error(self, mock_module_with_list_func):
         func_path = "mock_func_path"
         with pytest.raises(
-            DynamicListError,
-            match="Unable to display list of items due to 'Failed to parse function from function path: "
+            RetrieveToolFuncResultError,
+            match="Unable to retrieve result due to 'Failed to parse function from function path: "
             "'mock_func_path'. Expected format: format 'my_module.my_func'. Detailed error: not enough "
             "values to unpack \\(expected 2, got 1\\)'. \nPlease contact the tool author/support team for "
             "troubleshooting assistance.",
@@ -352,8 +352,8 @@ class TestToolUtils:
 
         func_path = "fake_tool_pkg.tools.tool_with_dynamic_list_input.my_list_func"
         with pytest.raises(
-            DynamicListError,
-            match="Unable to display list of items due to 'Failed to parse function from function path: "
+            RetrieveToolFuncResultError,
+            match="Unable to retrieve result due to 'Failed to parse function from function path: "
             "'fake_tool_pkg.tools.tool_with_dynamic_list_input.my_list_func'. Expected format: format "
             "'my_module.my_func'. Detailed error: No module named 'fake_tool_pkg''. \nPlease contact the tool "
             "author/support team for troubleshooting assistance.",
@@ -362,10 +362,10 @@ class TestToolUtils:
 
         func_path = "my_tool_package.tools.tool_with_dynamic_list_input.my_field"
         with pytest.raises(
-            DynamicListError,
-            match="Unable to display list of items due to 'Failed to parse function from function path: "
+            RetrieveToolFuncResultError,
+            match="Unable to retrieve result due to 'Failed to parse function from function path: "
             "'my_tool_package.tools.tool_with_dynamic_list_input.my_field'. Expected format: "
-            "format 'my_module.my_func'. Detailed error: Unable to display list of items due to ''1' "
+            "format 'my_module.my_func'. Detailed error: Unable to retrieve result due to ''1' "
             "is not callable.'. \nPlease contact the tool author/support team for troubleshooting assistance.",
         ):
             load_function_from_function_path(func_path)
@@ -389,7 +389,7 @@ class TestToolUtils:
     )
     def test_validate_tool_func_result(self, func_call_scenario, result, err_msg):
         error_message = (
-            f"Unable to retrieve tool func result due to '{err_msg}'. \nPlease contact the tool author/support team "
+            f"Unable to retrieve result due to '{err_msg}'. \nPlease contact the tool author/support team "
             f"for troubleshooting assistance."
         )
         with pytest.raises(RetrieveToolFuncResultValidationError) as e: